pyproject.toml
[project]
name = "study-thing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.39.0",
    "marker-pdf>=1.8.0",
    "markitdown[pdf]>=0.1.2",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "pymupdf4llm>=0.0.26",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=9.4.0",
]
